+++ ts/features/barcode/screens/BarcodeScanScreen.tsx
@@ -1,14 +1,15 @@
 import {
   Divider,
+  IOToast,
   ListItemNav,
-  VSpacer,
-  IOToast
+  VSpacer
 } from "@pagopa/io-app-design-system";
 import { useNavigation } from "@react-navigation/native";
 import { Alert, View } from "react-native";
 import ReactNativeHapticFeedback, {
   HapticFeedbackTypes
 } from "react-native-haptic-feedback";
+import { useHardwareBackButton } from "../../../hooks/useHardwareBackButton";
 import { useOpenDeepLink } from "../../../hooks/useOpenDeepLink";
 import I18n from "../../../i18n";
 import { mixpanelTrack } from "../../../mixpanel";
@@ -17,13 +18,22 @@ import {
   IOStackNavigationProp
 } from "../../../navigation/params/AppParamsList";
 import { useIOSelector } from "../../../store/hooks";
-import { barcodesScannerConfigSelector } from "../../../store/reducers/backendStatus/remoteConfig";
+import {
+  barcodesScannerConfigSelector,
+  isPnRemoteEnabledSelector
+} from "../../../store/reducers/backendStatus/remoteConfig";
+import { isIdPayLocallyEnabledSelector } from "../../../store/reducers/persistedPreferences.ts";
 import { emptyContextualHelp } from "../../../utils/emptyContextualHelp";
 import { useIOBottomSheetModal } from "../../../utils/hooks/bottomSheet";
+import { FCI_ROUTES } from "../../fci/navigation/routes";
 import { IdPayPaymentRoutes } from "../../idpay/payment/navigation/routes";
+import { ITW_REMOTE_ROUTES } from "../../itwallet/presentation/remote/navigation/routes.ts";
+import { PaymentsBarcodeRoutes } from "../../payments/barcode/navigation/routes";
+import { usePagoPaPayment } from "../../payments/checkout/hooks/usePagoPaPayment";
 import { PaymentsCheckoutRoutes } from "../../payments/checkout/navigation/routes";
-import * as analytics from "../analytics";
+import { paymentAnalyticsDataSelector } from "../../payments/history/store/selectors";
 import * as paymentsAnalytics from "../../payments/home/analytics";
+import * as analytics from "../analytics";
 import { BarcodeScanBaseScreenComponent } from "../components/BarcodeScanBaseScreenComponent";
 import { useIOBarcodeFileReader } from "../hooks/useIOBarcodeFileReader";
 import {
@@ -37,19 +47,13 @@ import {
 } from "../types/IOBarcode";
 import { BarcodeFailure } from "../types/failure";
 import { getIOBarcodesByType } from "../utils/getBarcodesByType";
-import { PaymentsBarcodeRoutes } from "../../payments/barcode/navigation/routes";
-import { useHardwareBackButton } from "../../../hooks/useHardwareBackButton";
-import { usePagoPaPayment } from "../../payments/checkout/hooks/usePagoPaPayment";
-import { FCI_ROUTES } from "../../fci/navigation/routes";
-import { paymentAnalyticsDataSelector } from "../../payments/history/store/selectors";
-import { isIdPayLocallyEnabledSelector } from "../../../store/reducers/persistedPreferences.ts";
-import { ITW_REMOTE_ROUTES } from "../../itwallet/presentation/remote/navigation/routes.ts";
 
 const BarcodeScanScreen = () => {
   const navigation = useNavigation<IOStackNavigationProp<AppParamsList>>();
   const openDeepLink = useOpenDeepLink();
   const isIdPayEnabled = useIOSelector(isIdPayLocallyEnabledSelector);
   const paymentAnalyticsData = useIOSelector(paymentAnalyticsDataSelector);
+  const isSendEnabled = useIOSelector(isPnRemoteEnabledSelector);
 
   const { startPaymentFlowWithRptId } = usePagoPaPayment();
 
@@ -67,7 +71,7 @@ const BarcodeScanScreen = () => {
   );
 
   const barcodeTypes: Array<IOBarcodeType> = IO_BARCODE_ALL_TYPES.filter(type =>
-    type === "IDPAY" ? isIdPayEnabled : true
+    type === "IDPAY" ? isIdPayEnabled : type === "SEND" ? isSendEnabled : true
   );
 
   /**
@@ -158,6 +162,25 @@ const BarcodeScanScreen = () => {
         navigation.navigate(ITW_REMOTE_ROUTES.MAIN, {
           screen: ITW_REMOTE_ROUTES.REQUEST_VALIDATION,
           params: barcode.itwRemoteRequestPayload
+        });
+        break;
+      case "SEND":
+        // TODO navigate to AAR screen/flow. It is not
+        // implemented yet so we keep treating it as an error
+        //
+        // navigation.navigate(MESSAGES_ROUTES.MESSAGES_NAVIGATOR, {
+        //   screen: PN_ROUTES.MAIN,
+        //   params: {
+        //     screen: PN_ROUTES.QR_SCAN_FLOW,
+        //     params: {
+        //       aarUrl: barcode.qrCodeContent
+        //     }
+        //   }
+        // });
+        handleBarcodeError({
+          content: "",
+          format: "QR_CODE",
+          reason: "UNKNOWN_CONTENT"
         });
         break;
     }