+++ ts/features/idpay/onboarding/machine/actions.ts
@@ -2,6 +2,7 @@ import * as O from "fp-ts/lib/Option";
 import { useIONavigation } from "../../../../navigation/params/AppParamsList";
 import { useIODispatch } from "../../../../store/hooks";
 import { refreshSessionToken } from "../../../authentication/fastLogin/store/actions/tokenRefreshActions";
+import { SERVICES_ROUTES } from "../../../services/common/navigation/routes";
 import { IDPayDetailsRoutes } from "../../details/navigation";
 import { IdPayOnboardingRoutes } from "../navigation/routes";
 import * as Context from "./context";
@@ -64,7 +65,20 @@ export const createActionsImplementation = (
       screen: IdPayOnboardingRoutes.IDPAY_ONBOARDING_ENABLE_NOTIFICATIONS
     });
 
+  const navigateToEnableMessageScreen = () =>
+    navigation.navigate(IdPayOnboardingRoutes.IDPAY_ONBOARDING_MAIN, {
+      screen: IdPayOnboardingRoutes.IDPAY_ONBOARDING_ENABLE_MESSAGE
+    });
+
   const closeOnboarding = () => navigation.popToTop();
+
+  const closeOnboardingSuccess = () =>
+    navigation.navigate(SERVICES_ROUTES.SERVICES_HOME);
+
+  const navigateToLoadingScreen = () =>
+    navigation.navigate(IdPayOnboardingRoutes.IDPAY_ONBOARDING_MAIN, {
+      screen: IdPayOnboardingRoutes.IDPAY_ONBOARDING_LOADING
+    });
 
   const handleSessionExpired = () => {
     dispatch(
@@ -86,7 +100,10 @@ export const createActionsImplementation = (
     navigateToInitiativeMonitoringScreen,
     navigateToInputFormScreen,
     navigateToEnableNotificationScreen,
+    navigateToEnableMessageScreen,
+    navigateToLoadingScreen,
     closeOnboarding,
+    closeOnboardingSuccess,
     handleSessionExpired
   };
 };