+++ ts/features/idpay/onboarding/machine/machine.ts
@@ -34,9 +34,12 @@ export const idPayOnboardingMachine = setup({
     navigateToFailureScreen: notImplementedStub,
     navigateToInitiativeMonitoringScreen: notImplementedStub,
     closeOnboarding: notImplementedStub,
+    closeOnboardingSuccess: notImplementedStub,
     handleSessionExpired: notImplementedStub,
     navigateToInputFormScreen: notImplementedStub,
-    navigateToEnableNotificationScreen: notImplementedStub
+    navigateToEnableNotificationScreen: notImplementedStub,
+    navigateToEnableMessageScreen: notImplementedStub,
+    navigateToLoadingScreen: notImplementedStub
   },
   actors: {
     getInitiativeInfo: fromPromise<InitiativeDataDTO, string>(
@@ -89,7 +92,8 @@ export const idPayOnboardingMachine = setup({
     isSessionExpired: ({ event }: { event: IdPayOnboardingEvents }) =>
       "error" in event && event.error === InitiativeFailureType.SESSION_EXPIRED,
     shouldShowEnableNotificationOnClose: ({ context }) =>
-      !context.isPushNotificationsEnabled
+      !context.isPushNotificationsEnabled,
+    hasMessageConsent: ({ context }) => !context.hasInbox
   }
 }).createMachine({
   id: "idpay-onboarding",
@@ -107,7 +111,9 @@ export const idPayOnboardingMachine = setup({
         "start-onboarding": {
           guard: "assertServiceId",
           actions: assign(({ event }) => ({
-            serviceId: event.serviceId
+            serviceId: event.serviceId,
+            hasInbox: event.hasInbox,
+            currentStep: 1
           })),
           target: "LoadingInitiative"
         }
@@ -175,6 +181,25 @@ export const idPayOnboardingMachine = setup({
     },
 
     DisplayingInitiativeInfo: {
+      entry: "navigateToInitiativeDetailsScreen",
+      actions: assign(() => ({
+        currentStep: 1
+      })),
+      on: {
+        next: [
+          {
+            guard: "hasMessageConsent",
+            target: "EnableMessage"
+          },
+          {
+            target: "AcceptingTos"
+          }
+        ]
+      }
+    },
+
+    EnableMessage: {
+      entry: "navigateToEnableMessageScreen",
       on: {
         next: {
           target: "AcceptingTos"
@@ -252,10 +277,16 @@ export const idPayOnboardingMachine = setup({
       on: {
         next: [
           {
+            actions: assign(({ context }) => ({
+              currentStep: context.currentStep + 1
+            })),
             guard: "hasSelfDecalrationList",
             target: "DisplayingSelfDeclarationList"
           },
           {
+            actions: assign(({ context }) => ({
+              currentStep: context.currentStep + 1
+            })),
             target: "AcceptingCriteria"
           }
         ],
@@ -291,10 +322,16 @@ export const idPayOnboardingMachine = setup({
           on: {
             back: [
               {
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep - 1
+                })),
                 guard: "hasPdndCriteria",
                 target: "#idpay-onboarding.DisplayingPdndCriteria"
               },
               {
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep - 1
+                })),
                 target: "#idpay-onboarding.DisplayingInitiativeInfo"
               }
             ],
@@ -309,7 +346,10 @@ export const idPayOnboardingMachine = setup({
             next: [
               {
                 guard: "hasMultiSelfDeclarationList",
-                target: "DisplayingMultiSelfDeclarationList"
+                target: "DisplayingMultiSelfDeclarationList",
+                actions: assign(({ context }) => ({
+                  currentStep: context.currentStep + 1
+                }))
               },
               {
                 guard: "hasInputFormDeclaration",
@@ -330,6 +370,7 @@ export const idPayOnboardingMachine = setup({
               on: {
                 "select-multi-consent": {
                   actions: assign(({ context, event }) => ({
+                    currentStep: context.currentStep + 1,
                     selfDeclarationsMultiAnwsers: {
                       ...context.selfDeclarationsMultiAnwsers,
                       [context.selfDeclarationsMultiPage]: event.data
@@ -344,20 +385,35 @@ export const idPayOnboardingMachine = setup({
                       "hasBooleanSelfDeclarationList"
                     ]),
                     target:
-                      "#idpay-onboarding.DisplayingSelfDeclarationList.DisplayingBooleanSelfDeclarationList"
+                      "#idpay-onboarding.DisplayingSelfDeclarationList.DisplayingBooleanSelfDeclarationList",
+                    actions: assign(({ context }) => ({
+                      selfDeclarationsMultiPage: Math.max(
+                        0,
+                        +context.selfDeclarationsMultiPage - 1
+                      ),
+                      currentStep: context.currentStep - 1
+                    }))
                   },
                   {
                     guard: and(["isFirstMultiConsentPage", "hasPdndCriteria"]),
-                    target: "#idpay-onboarding.DisplayingPdndCriteria"
+                    target: "#idpay-onboarding.DisplayingPdndCriteria",
+                    actions: assign(({ context }) => ({
+                      selfDeclarationsMultiPage: Math.max(
+                        0,
+                        +context.selfDeclarationsMultiPage - 1
+                      ),
+                      currentStep: context.currentStep - 1
+                    }))
                   },
                   {
                     guard: "isFirstMultiConsentPage",
-                    target: "#idpay-onboarding.DisplayingInitiativeInfo"
-                  },
-                  {
+                    target: "#idpay-onboarding.DisplayingInitiativeInfo",
                     actions: assign(({ context }) => ({
-                      selfDeclarationsMultiPage:
+                      selfDeclarationsMultiPage: Math.max(
+                        0,
                         +context.selfDeclarationsMultiPage - 1
+                      ),
+                      currentStep: context.currentStep - 1
                     }))
                   }
                 ]
@@ -377,7 +433,8 @@ export const idPayOnboardingMachine = setup({
                 {
                   actions: assign(({ context }) => ({
                     selfDeclarationsMultiPage:
-                      +context.selfDeclarationsMultiPage + 1
+                      +context.selfDeclarationsMultiPage + 1,
+                    currentStep: context.currentStep + 1
                   })),
                   target: "DisplayingMultiSelfDeclarationItem"
                 }
@@ -397,7 +454,8 @@ export const idPayOnboardingMachine = setup({
                     selfDeclarationsTextAnswers: {
                       ...context.selfDeclarationsTextAnswers,
                       [context.activeTextConsentPage]: event.criteria
-                    }
+                    },
+                    currentStep: context.currentStep + 1
                   })),
                   target: "EvaluatingInputForm"
                 },
@@ -428,7 +486,8 @@ export const idPayOnboardingMachine = setup({
                   },
                   {
                     actions: assign(({ context }) => ({
-                      activeTextConsentPage: +context.activeTextConsentPage - 1
+                      activeTextConsentPage: +context.activeTextConsentPage - 1,
+                      currentStep: context.currentStep - 1
                     }))
                   }
                 ]
@@ -454,6 +513,7 @@ export const idPayOnboardingMachine = setup({
     },
 
     AcceptingCriteria: {
+      entry: "navigateToLoadingScreen",
       tags: [IdPayTags.Loading],
       invoke: {
         src: "acceptRequiredCriteria",
@@ -489,7 +549,7 @@ export const idPayOnboardingMachine = setup({
           target: "Idle"
         },
         close: {
-          actions: "closeOnboarding",
+          actions: "closeOnboardingSuccess",
           target: "Idle"
         }
       }
@@ -509,7 +569,7 @@ export const idPayOnboardingMachine = setup({
             guard: "shouldShowEnableNotificationOnClose"
           },
           {
-            actions: "closeOnboarding"
+            actions: "closeOnboardingSuccess"
           }
         ]
       }
@@ -518,7 +578,7 @@ export const idPayOnboardingMachine = setup({
     OnboardingFailure: {
       entry: "navigateToFailureScreen",
       on: {
-        next: {
+        "check-details": {
           actions: "navigateToInitiativeMonitoringScreen"
         }
       }