+++ ts/features/itwallet/identification/cie/screens/ItwCiePinScreen.tsx
@@ -9,7 +9,13 @@ import { Millisecond } from "@pagopa/ts-commons/lib/units";
 import { useHeaderHeight } from "@react-navigation/elements";
 import { useFocusEffect, useIsFocused } from "@react-navigation/native";
 import { useCallback, useEffect, useRef, useState } from "react";
-import { Keyboard, KeyboardAvoidingView, Platform, View } from "react-native";
+import {
+  Keyboard,
+  KeyboardAvoidingView,
+  Platform,
+  ScrollView,
+  View
+} from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { ContextualHelpPropsMarkdown } from "../../../../../components/screens/BaseScreenComponent";
 import { useHeaderSecondLevel } from "../../../../../hooks/useHeaderSecondLevel";
@@ -23,8 +29,9 @@ import {
   trackItWalletCiePinEnter,
   trackItWalletCiePinInfo
 } from "../../../analytics";
-import { ItwEidIssuanceMachineContext } from "../../../machine/provider";
-import { useCieInfoBottomSheet } from "../../hooks/useCieInfoBottomSheet";
+import { ItwEidIssuanceMachineContext } from "../../../machine/eid/provider";
+import { isL3FeaturesEnabledSelector } from "../../../machine/eid/selectors";
+import { useCieInfoBottomSheet } from "../hooks/useCieInfoBottomSheet";
 
 const CIE_PIN_LENGTH = 8;
 
@@ -38,6 +45,10 @@ export const ItwCiePinScreen = () => {
 
   const useCieUat = useIOSelector(isCieLoginUatEnabledSelector);
   const machineRef = ItwEidIssuanceMachineContext.useActorRef();
+  const isL3Enabled = ItwEidIssuanceMachineContext.useSelector(
+    isL3FeaturesEnabledSelector
+  );
+  const itw_flow = isL3Enabled ? "L3" : "L2";
 
   const [pin, setPin] = useState("");
   const pinPadViewRef = useRef<View>(null);
@@ -59,9 +70,9 @@ export const ItwCiePinScreen = () => {
 
   useFocusEffect(
     useCallback(() => {
-      trackItWalletCiePinEnter();
+      trackItWalletCiePinEnter(itw_flow);
       setAccessibilityFocus(pinPadViewRef, 300 as Millisecond);
-    }, [])
+    }, [itw_flow])
   );
 
   useHeaderSecondLevel({
@@ -90,43 +101,45 @@ export const ItwCiePinScreen = () => {
         style={{ flex: 1 }}
         keyboardVerticalOffset={headerHeight}
       >
-        <ContentWrapper>
-          <H2>
-            {I18n.t("features.itWallet.identification.cie.inputPin.title")}
-          </H2>
-          <VSpacer size={8} />
-          <IOButton
-            variant="link"
-            label={I18n.t(
-              "features.itWallet.identification.cie.inputPin.buttonLink"
-            )}
-            accessibilityLabel={I18n.t(
-              "features.itWallet.identification.cie.inputPin.buttonLink"
-            )}
-            onPress={() => {
-              trackItWalletCiePinInfo();
-              pinInfoBottomSheet.present();
-            }}
-          />
-          <VSpacer size={24} />
-          <View style={{ flex: 1 }}>
-            <OTPInput
-              ref={pinPadViewRef}
-              secret
-              value={pin}
+        <ScrollView contentContainerStyle={{ flexGrow: 1 }}>
+          <ContentWrapper>
+            <H2>
+              {I18n.t("features.itWallet.identification.cie.inputPin.title")}
+            </H2>
+            <VSpacer size={8} />
+            <IOButton
+              variant="link"
+              label={I18n.t(
+                "features.itWallet.identification.cie.inputPin.buttonLink"
+              )}
               accessibilityLabel={I18n.t(
-                "authentication.cie.pin.accessibility.label"
+                "features.itWallet.identification.cie.inputPin.buttonLink"
               )}
-              accessibilityHint={I18n.t(
-                "authentication.cie.pin.accessibility.hint"
-              )}
-              onValueChange={onPinChanged}
-              length={CIE_PIN_LENGTH}
-              autoFocus={isFocused}
-              key={isFocused ? "focused" : "unfocused"}
+              onPress={() => {
+                trackItWalletCiePinInfo(itw_flow);
+                pinInfoBottomSheet.present();
+              }}
             />
-          </View>
-        </ContentWrapper>
+            <VSpacer size={24} />
+            <View style={{ flex: 1 }}>
+              <OTPInput
+                ref={pinPadViewRef}
+                secret
+                value={pin}
+                accessibilityLabel={I18n.t(
+                  "authentication.cie.pin.accessibility.label"
+                )}
+                accessibilityHint={I18n.t(
+                  "authentication.cie.pin.accessibility.hint"
+                )}
+                onValueChange={onPinChanged}
+                length={CIE_PIN_LENGTH}
+                autoFocus={isFocused}
+                key={isFocused ? "focused" : "unfocused"}
+              />
+            </View>
+          </ContentWrapper>
+        </ScrollView>
       </KeyboardAvoidingView>
       {pinInfoBottomSheet.bottomSheet}
     </SafeAreaView>