+++ ts/features/pn/navigation/params.ts
@@ -1,3 +1,4 @@
+import { SendQRScanFlowProps } from "../qrCodeScan/screens/QrScanFlow";
 import { MessageAttachmentScreenRouteParams } from "../screens/MessageAttachmentScreen";
 import { MessageDetailsScreenRouteParams } from "../screens/MessageDetailsScreen";
 import { PaidPaymentScreenRouteParams } from "../screens/PaidPaymentScreen";
@@ -7,5 +8,9 @@ export type PnParamsList = {
   [PN_ROUTES.MESSAGE_DETAILS]: MessageDetailsScreenRouteParams;
   [PN_ROUTES.MESSAGE_ATTACHMENT]: MessageAttachmentScreenRouteParams;
   [PN_ROUTES.CANCELLED_MESSAGE_PAID_PAYMENT]: PaidPaymentScreenRouteParams;
+  // ENGAGEMENT BANNER
   [PN_ROUTES.ACTIVATION_BANNER_FLOW]: undefined;
+  // QR SCAN
+  [PN_ROUTES.QR_SCAN_FLOW]: SendQRScanFlowProps;
+  [PN_ROUTES.ENGAGEMENT_SCREEN]: undefined;
 };